+++ contracts/access/AccessControl.sol
@@ -92,7 +92,7 @@ abstract contract AccessControl is Context {
      * @dev Returns the number of accounts that have `role`. Can be used
      * together with {getRoleMember} to enumerate all bearers of a role.
      */
-    function getRoleMemberCount(bytes32 role) public view returns (uint256) {
+    function getRoleMemberCount(bytes32 role) public view returns (uint) {
         return _roles[role].members.length();
     }
 
@@ -108,7 +108,7 @@ abstract contract AccessControl is Context {
      * https://forum.openzeppelin.com/t/iterating-over-elements-on-enumerableset-in-openzeppelin-contracts/2296[forum post]
      * for more information.
      */
-    function getRoleMember(bytes32 role, uint256 index) public view returns (address) {
+    function getRoleMember(bytes32 role, uint index) public view returns (address) {
         return _roles[role].members.at(index);
     }
