+++ contracts/crowdsale/price/IncreasingPriceCrowdsale.sol
@@ -10,17 +10,17 @@ import "../../math/SafeMath.sol";
  * the amount of tokens per wei contributed. Thus, the initial rate must be greater than the final rate.
  */
 contract IncreasingPriceCrowdsale is TimedCrowdsale {
-    using SafeMath for uint256;
+    using SafeMath for uint;
 
-    uint256 private _initialRate;
-    uint256 private _finalRate;
+    uint private _initialRate;
+    uint private _finalRate;
 
     /**
      * @dev Constructor, takes initial and final rates of tokens received per wei contributed.
      * @param initialRate Number of tokens a buyer gets per wei at the start of the crowdsale
      * @param finalRate Number of tokens a buyer gets per wei at the end of the crowdsale
      */
-    constructor (uint256 initialRate, uint256 finalRate) public {
+    constructor (uint initialRate, uint finalRate) public {
         require(finalRate > 0, "IncreasingPriceCrowdsale: final rate is 0");
         // solhint-disable-next-line max-line-length
         require(initialRate > finalRate, "IncreasingPriceCrowdsale: initial rate is not greater than final rate");
@@ -32,21 +32,21 @@ contract IncreasingPriceCrowdsale is TimedCrowdsale {
      * The base rate function is overridden to revert, since this crowdsale doesn't use it, and
      * all calls to it are a mistake.
      */
-    function rate() public view returns (uint256) {
+    function rate() public view returns (uint) {
         revert("IncreasingPriceCrowdsale: rate() called");
     }
 
     /**
      * @return the initial rate of the crowdsale.
      */
-    function initialRate() public view returns (uint256) {
+    function initialRate() public view returns (uint) {
         return _initialRate;
     }
 
     /**
      * @return the final rate of the crowdsale.
      */
-    function finalRate() public view returns (uint256) {
+    function finalRate() public view returns (uint) {
         return _finalRate;
     }
 
@@ -55,15 +55,15 @@ contract IncreasingPriceCrowdsale is TimedCrowdsale {
      * Note that, as price _increases_ with time, the rate _decreases_.
      * @return The number of tokens a buyer gets per wei at a given time
      */
-    function getCurrentRate() public view returns (uint256) {
+    function getCurrentRate() public view returns (uint) {
         if (!isOpen()) {
             return 0;
         }
 
         // solhint-disable-next-line not-rely-on-time
-        uint256 elapsedTime = block.timestamp.sub(openingTime());
-        uint256 timeRange = closingTime().sub(openingTime());
-        uint256 rateRange = _initialRate.sub(_finalRate);
+        uint elapsedTime = block.timestamp.sub(openingTime());
+        uint timeRange = closingTime().sub(openingTime());
+        uint rateRange = _initialRate.sub(_finalRate);
         return _initialRate.sub(elapsedTime.mul(rateRange).div(timeRange));
     }
 
@@ -72,8 +72,8 @@ contract IncreasingPriceCrowdsale is TimedCrowdsale {
      * @param weiAmount The value in wei to be converted into tokens
      * @return The number of tokens _weiAmount wei will buy at present time
      */
-    function _getTokenAmount(uint256 weiAmount) internal view returns (uint256) {
-        uint256 currentRate = getCurrentRate();
+    function _getTokenAmount(uint weiAmount) internal view returns (uint) {
+        uint currentRate = getCurrentRate();
         return currentRate.mul(weiAmount);
     }
 }
