+++ contracts/mocks/ContextMock.sol
@@ -11,9 +11,9 @@ contract ContextMock is Context {
         emit Sender(_msgSender());
     }
 
-    event Data(bytes data, uint256 integerValue, string stringValue);
+    event Data(bytes data, uint integerValue, string stringValue);
 
-    function msgData(uint256 integerValue, string memory stringValue) public {
+    function msgData(uint integerValue, string memory stringValue) public {
         emit Data(_msgData(), integerValue, stringValue);
     }
 }
@@ -23,7 +23,7 @@ contract ContextMockCaller {
         context.msgSender();
     }
 
-    function callData(ContextMock context, uint256 integerValue, string memory stringValue) public {
+    function callData(ContextMock context, uint integerValue, string memory stringValue) public {
         context.msgData(integerValue, stringValue);
     }
 }
