+++ contracts/mocks/ERC20Mock.sol
@@ -10,24 +10,24 @@ contract ERC20Mock is ERC20 {
         string memory name,
         string memory symbol,
         address initialAccount,
-        uint256 initialBalance
+        uint initialBalance
     ) public payable ERC20(name, symbol) {
         _mint(initialAccount, initialBalance);
     }
 
-    function mint(address account, uint256 amount) public {
+    function mint(address account, uint amount) public {
         _mint(account, amount);
     }
 
-    function burn(address account, uint256 amount) public {
+    function burn(address account, uint amount) public {
         _burn(account, amount);
     }
 
-    function transferInternal(address from, address to, uint256 value) public {
+    function transferInternal(address from, address to, uint value) public {
         _transfer(from, to, value);
     }
 
-    function approveInternal(address owner, address spender, uint256 value) public {
+    function approveInternal(address owner, address spender, uint value) public {
         _approve(owner, spender, value);
     }
 }
