+++ contracts/mocks/ERC20PausableMock.sol
@@ -10,7 +10,7 @@ contract ERC20PausableMock is ERC20Pausable {
         string memory name,
         string memory symbol,
         address initialAccount,
-        uint256 initialBalance
+        uint initialBalance
     ) public ERC20(name, symbol) {
         _mint(initialAccount, initialBalance);
     }
@@ -23,11 +23,11 @@ contract ERC20PausableMock is ERC20Pausable {
         _unpause();
     }
 
-    function mint(address to, uint256 amount) public {
+    function mint(address to, uint amount) public {
         _mint(to, amount);
     }
 
-    function burn(address from, uint256 amount) public {
+    function burn(address from, uint amount) public {
         _burn(from, amount);
     }
 }
