+++ contracts/mocks/ERC20PermitMock.sol
@@ -9,12 +9,12 @@ contract ERC20PermitMock is ERC20Permit {
         string memory name,
         string memory symbol,
         address initialAccount,
-        uint256 initialBalance
+        uint initialBalance
     ) public payable ERC20(name, symbol) ERC20Permit(name) {
         _mint(initialAccount, initialBalance);
     }
 
-    function getChainId() external view returns (uint256 chainId) {
+    function getChainId() external view returns (uint chainId) {
         this; // silence state mutability warning without generating bytecode - see https://github.com/ethereum/solidity/issues/2691
         // solhint-disable-next-line no-inline-assembly
         assembly {
